requirements.txt
Django==2.1.5
pytz==2018.9
djangorestframework==3.9.0
Flask==1.0.2
requests==2.20.0
